Fix ambiguity build error.

+++ src/rpcblockchain.cpp
@@ -440,8 +440,8 @@ Object blockToJSON(const CBlock& block, const CBlockIndex* blockindex, bool fPri
         std::string superblock=UnpackBinarySuperblock(bb.superblock);
         std::string neural_hash = GetQuorumHash(superblock);
         result.push_back(Pair("SuperblockHash", neural_hash));
-        result.push_back(Pair("SuperblockUnpackedLength", (long)superblock.length()));
-        result.push_back(Pair("SuperblockLength", (long)bb.superblock.length()));
+        result.push_back(Pair("SuperblockUnpackedLength", (int)superblock.length()));
+        result.push_back(Pair("SuperblockLength", (int)bb.superblock.length()));
         bool bIsBinary = Contains(bb.superblock,"<BINARY>");
         result.push_back(Pair("IsBinary",bIsBinary));
         if(fPrintTransactionDetail)
